+++ textblob_de/compat.py
@@ -1,4 +1,11 @@
 # -*- coding: utf-8 -*-
+'''Code imported from ``textblob-fr`` sample extension.
+
+:repo: `https://github.com/sloria/textblob-fr`_
+:source: textblob_fr/compat.py
+:version: 2013-10-28 (5c6329d209)
+
+'''
 import sys
 
 PY2 = int(sys.version[0]) == 2