+++ textblob_de/sentiments.py
@@ -1,26 +1,40 @@
 # -*- coding: utf-8 -*-
 """German sentiment analysis implementations.
 
+Code adapted from ``textblob-fr`` sample extension.
+
+:repo: `https://github.com/sloria/textblob-fr`_
+:source: textblob_fr/sentiments.py
+:version: 2013-10-28 (a88e86a76a)
+
+:modified: July 2014 <m.killer@langui.ch>
+
+Main resource for ``de-sentiment.xml``:
+
+* `German Polarity Lexicon <http://bics.sentimental.li/index.php/downloads>`_
+* See xml comment section in ``de-sentiment.xml`` for details
+
 .. todo::
 
-       extract/compute/compile German Subjectivity Lexicon from
-       publicly available resources.
+       enhance German Polarity Lexicon, using publicly available resources.
+       
+       Missing values:
+       
+       * Subjectivity
+       * (Intensity)
        
        Possible sources:
        
-       * Wait for ``de-sentiment.xml`` to be added to `pattern-de`_
-       * `German Polarity Lexicon`_
-       * `IGGSA`_ (Interest Group on German Sentiment Analysis)
-       * `GermanPolarityClues`_ - A Lexical Resource for German Sentiment Analysis
+       * `Pattern Projext <http://www.clips.ua.ac.be/pages/pattern>`_
+       
+           * fr-sentiment.xml / en-sentiment.xml / nl-sentiment.xml
+           
+       * `IGGSA <https://sites.google.com/site/iggsahome/>`_
 
-
-.. _pattern-de: http://www.clips.ua.ac.be/pages/pattern-de
-.. _German Polarity Lexicon: http://bics.sentimental.li/index.php/downloads/
-.. _IGGSA: https://sites.google.com/site/iggsahome/
-.. _GermanPolarityClues: http://www.ulliwaltinger.de/sentiment/
 """
 from __future__ import absolute_import
 from textblob.base import BaseSentimentAnalyzer, CONTINUOUS
+from textblob_de.tokenizers import get_arg_tokenizer
 from textblob_de.de import sentiment as pattern_sentiment
 
 
@@ -38,5 +52,4 @@ class PatternAnalyzer(BaseSentimentAnalyzer):
         """Return the sentiment as a tuple of the form:
         ``(polarity, subjectivity)``
         """
-        #raise NotImplementedError
         return pattern_sentiment(text)